tools/serial_monitor/monitor_tool/requirements.txt
pyserial
matplotlib
numpy
openai
pyinstaller
pandas
scikit-learn
joblib